+++ src/main/java/ch/business/quickline/service/MitarbeiterService.java
@@ -11,8 +11,8 @@ public interface MitarbeiterService {
 	public List<Mitarbeiter> mitarbeiterMasterBewertungRangliste();
 	public List<Mitarbeiter> mitarbeiterSelbstBewertungRangliste();
 	public Mitarbeiter findByMitarbeiterId(Integer id);
-	public BigDecimal retrieveMasterBewertungDurchschnitt(Integer id);
-	public BigDecimal retrieveSelbstBewertungDurchschnitt(Integer id);
+	public Double retrieveMasterBewertungDurchschnitt(Integer id);
+	public Double retrieveSelbstBewertungDurchschnitt(Integer id);
 	public long count ();
