+++ src/main/java/ch/business/quickline/service/SkillService.java
@@ -9,8 +9,8 @@ public interface SkillService {
 	public Skill save (Skill skill);
 	public List<Skill>findAll();
 	public Skill findBySkillId(Integer id);
-	public BigDecimal retrieveMasterBewertungDurchschnitt(Integer id);
-	public BigDecimal retrieveSelbstBewertungDurchschnitt(Integer id); 
+	public Double retrieveMasterBewertungDurchschnitt(Integer id);
+	public Double retrieveSelbstBewertungDurchschnitt(Integer id); 
 	public long count();
 
 }
