+++ src/main/java/ch/business/quickline/service/SkillServiceImpl.java
@@ -33,14 +33,14 @@ public class SkillServiceImpl implements SkillService {
 	}
 	
 	
-	public BigDecimal retrieveMasterBewertungDurchschnitt(Integer id){
-    	BigDecimal average = findBySkillId(id).getSkillMasterBewertungDurchschnitt();
+	public Double retrieveMasterBewertungDurchschnitt(Integer id){
+    	Double average = findBySkillId(id).getSkillMasterBewertungDurchschnitt();
     	
     	return average;
 }
 
-    public BigDecimal retrieveSelbstBewertungDurchschnitt(Integer id){
-    	BigDecimal average = findBySkillId(id).getSkillSelbstBewertungDurchschnitt();
+    public Double retrieveSelbstBewertungDurchschnitt(Integer id){
+    	Double average = findBySkillId(id).getSkillSelbstBewertungDurchschnitt();
     	
     	return average;
 }
